requirements.txt
httpx
beautifulsoup4
feedparser
python-dotenv